reorganizing code and moving the map creation into the server components
